requirements.txt
annotated-types==0.7.0
anthropic==0.68.1
anyio==4.11.0
autogen-agentchat==0.7.4
autogen-core==0.7.4
autogen-ext==0.7.4
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
distro==1.9.0
docstring-parser==0.17.0
fastapi==0.118.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib-metadata==8.7.0
jinja2==3.1.6
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
langchain==0.3.27
langchain-core==0.3.76
langchain-text-splitters==0.3.11
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langsmith==0.4.31
markupsafe==3.0.3
opentelemetry-api==1.37.0
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
pillow==11.3.0
protobuf==5.29.5
pyautogen==0.10.0
pydantic==2.11.9
pydantic-core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
pyyaml==6.0.3
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
sqlalchemy==2.0.43
starlette==0.48.0
tenacity==9.1.2
tiktoken==0.11.0
typing-extensions==4.15.0
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.37.0
xxhash==3.5.0
zipp==3.23.0
zstandard==0.25.0

# Testing dependencies
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
pytest-mock==3.14.0
httpx==0.28.1
playwright==1.51.0
